Revert "ED-25 Implement api"
