--- Dockerfile
@@ -1,20 +1,0 @@
-FROM python:3.10-slim
-
-WORKDIR /app
-
-COPY requirements.txt .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-RUN apt-get update
-RUN apt-get install -y wget
-RUN wget -q https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-RUN apt-get install -y ./google-chrome-stable_current_amd64.deb
-
-RUN apt-get install -y ffmpeg
-
-COPY . .
-
-EXPOSE 8000
-
-CMD ["python", "src/main.py"]
