--- requirements.txt
@@ -1,6 +1,0 @@
-fastapi
-uvicorn[standard]
-pydantic
-selenium
-yt-dlp
-aiohttp
