--- src/main.py
@@ -1,5 +1,0 @@
-import uvicorn
-from api.api import app
-
-if __name__ == "__main__":
-    uvicorn.run(app, host="0.0.0.0", port=8000)
